requirement.txt
elevenlabs
openai
python-dotenv
pillow
pyaudio
langchain-community
